refactor(surfaces): move query diagnostics and mutation result to shared payloads
Daemon HTTP no longer imports MCPQueryMissDiagnosticsPayload or
MutationResultPayload from polylogue.mcp.payloads. Both daemon list/
search diagnostics and the reset mutation handler now use the shared
payloads from polylogue/surfaces/payloads.py.

MutationResultPayload moved from MCP to shared surfaces; MCP re-exports
it for backward compatibility.
